chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
openai
jinja2
pydantic